# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: Tradosphere OS Gateway API
  version: "1.0.0"
  description: |-
    Unified API gateway for Tradosphere OS (Sprint 9, task 9.4). Implements
    the architecture decided in full in Decision D19 (EXECUTION_BOOK.md)
    before any gateway code was written.

    Two route families exist side by side:

    - Proxied (Auth, Market Data, Education, Portfolio, Analytics): the
      gateway reverse-proxies to each service's existing HTTP API via a
      hand-rolled wildcard handler that forwards the Authorization header
      verbatim. Each upstream service remains the sole auth authority for
      its own routes -- the gateway does NOT re-authenticate proxied
      requests.
    - In-process (Research, AI Council, CIO, Paper Trading, Journal): these
      five services are library-only with no HTTP surface of their own, so
      their routes are implemented directly on the gateway and authenticated
      by the gateway's own requireAuth (identical factory pattern to every
      other service in this repo).

    All domain routes are versioned under /v1/<service>/.... Infra routes
    (/health, /health/services, /metrics, /stream, /documentation,
    /openapi.yaml) are unversioned and, along with auth's own signup/login,
    are the only unauthenticated routes in the gateway.
servers:
  - url: http://localhost:4000
    description: Local gateway (apps/api)
...
security:
  - bearerAuth: []
paths:
...
  /v1/portfolio/positions:
    get:
      operationId: portfolioListPositions
      summary: Current open positions, netted per symbol
      description: |-
        Computed fresh from journal_entries on every call, never cached
        (Decision D17) -- closed trades no longer contribute, and a symbol
        whose open trades net to exactly zero is omitted rather than
        reported as a zero-quantity row.
      tags: [Portfolio]
      responses:
        '200':
          description: The caller's open positions, sorted by symbol.
          content:
            application/json:
              schema:
                type: object
                properties:
                  positions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Position'
                required: [positions]
        '401':
          description: Missing or invalid bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/portfolio/cash:
    get:
      operationId: portfolioGetCash
      summary: Current cash balance
      tags: [Portfolio]
      responses:
        '200':
          description: Starting cash plus every trade's signed cash effect.
          content:
            application/json:
              schema:
                type: object
                properties:
                  cashBalance:
                    type: number
                  startingCash:
                    type: number
                required: [cashBalance, startingCash]
        '401':
          description: Missing or invalid bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/portfolio/pnl:
    get:
      operationId: portfolioGetPnl
      summary: Realized and unrealized P&L
      tags: [Portfolio]
      responses:
        '200':
          description: |-
            P&L breakdown from a single mark-to-market pass. Open positions
            with no live tick are excluded from unrealizedPnl and listed in
            missingPriceSymbols rather than silently priced at zero.
          content:
            application/json:
              schema:
                type: object
                properties:
                  realizedPnl:
                    type: number
                  unrealizedPnl:
                    type: number
                  missingPriceSymbols:
                    type: array
                    items:
                      type: string
                required: [realizedPnl, unrealizedPnl, missingPriceSymbols]
        '401':
          description: Missing or invalid bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/portfolio/summary:
    get:
      operationId: portfolioGetSummary
      summary: Full portfolio summary (cash, positions, P&L, total equity)
      description: Every figure comes from one mark-to-market pass, so cash, positions, and P&L can never disagree about the price used for a given symbol within the same response.
      tags: [Portfolio]
      responses:
        '200':
          description: Portfolio summary for the caller.
          content:
            application/json:
              schema:
                type: object
                properties:
                  cashBalance:
                    type: number
                  positionsValue:
                    type: number
                  realizedPnl:
                    type: number
                  unrealizedPnl:
                    type: number
                  totalEquity:
                    type: number
                  positions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Position'
                  missingPriceSymbols:
                    type: array
                    items:
                      type: string
                required: [cashBalance, positionsValue, realizedPnl, unrealizedPnl, totalEquity, positions, missingPriceSymbols]
        '401':
          description: Missing or invalid bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/portfolio/snapshot:
    post:
      operationId: portfolioCreateSnapshot
      summary: Write a point-in-time MTM snapshot (Equity Curve / Portfolio History source row)
      description: |-
        Refused outright with 409 if any open position has no live price --
        a permanent history row is never written with a silently incomplete
        equity figure (Forge charter rule 2, Decision D17). Every numeric
        field is server-computed from journal_entries/market_ticks; the
        body may only supply a label and/or the as-of timestamp, which
        defaults to "now" when omitted. A request with no body at all is
        valid and treated the same as an empty object.
      tags: [Portfolio]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSnapshotInput'
      responses:
        '201':
          description: Snapshot written.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortfolioSnapshot'
        '400':
          description: Validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationFailure'
        '401':
          description: Missing or invalid bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: One or more open positions have no live price -- snapshot refused, nothing written.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  missingPriceSymbols:
                    type: array
                    items:
                      type: string
                required: [error, missingPriceSymbols]
  /v1/portfolio/history:
    get:
      operationId: portfolioGetHistory
      summary: Snapshot history (Equity Curve / Portfolio History)
      description: Serves both the "Equity Curve" and "Portfolio History" SPRINT_BOOK.md deliverables from one mechanism (Decision D17) -- the persisted rows written by POST /v1/portfolio/snapshot.
      tags: [Portfolio]
      parameters:
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: Inclusive lower bound on asOf. Omitted means no floor.
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: Inclusive upper bound on asOf. Omitted means no ceiling.
      responses:
        '200':
          description: Snapshots for the caller within the requested range.
          content:
            application/json:
              schema:
                type: object
                properties:
                  history:
                    type: array
                    items:
                      $ref: '#/components/schemas/PortfolioSnapshot'
                required: [history]
        '400':
          description: Validation failed (from/to not a valid ISO 8601 timestamp).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationFailure'
        '401':
          description: Missing or invalid bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/portfolio/performance:
    get:
      operationId: portfolioGetPerformance
      summary: Portfolio-level performance metrics (total return)
      description: |-
        Decision D17 scopes this to portfolio-level totals only -- win
        rate, risk/reward ratio, and drawdown need a trade-by-trade time
        series and belong to services/analytics instead.
      tags: [Portfolio]
      responses:
        '200':
          description: Performance metrics for the caller's account.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PerformanceMetrics'
        '401':
          description: Missing or invalid bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/portfolio/allocation:
    get:
      operationId: portfolioGetAllocation
      summary: Allocation / asset distribution across open positions
      description: allocationPct denominator is total absolute exposure across all positions, so a fully-hedged book (equal long and short) still sums to 100%, not 0%.
      tags: [Portfolio]
      responses:
        '200':
          description: Per-symbol allocation, sorted by share of exposure descending.
          content:
            application/json:
              schema:
                type: object
                properties:
                  allocation:
                    type: array
                    items:
                      $ref: '#/components/schemas/AllocationEntry'
                  missingPriceSymbols:
                    type: array
                    items:
                      type: string
                required: [allocation, missingPriceSymbols]
        '401':
          description: Missing or invalid bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/portfolio/risk:
    get:
      operationId: portfolioGetRisk
      summary: Portfolio-level risk exposure (gross/net/leverage)
      description: |-
        Decision D17 scopes this to portfolio-level exposure only --
        per-trade risk (stop-loss distance, R:R) depends on a CIO
        recommendation snapshot that is nullable on a journal entry
        (Decision D16), so that per-trade view is left to
        services/analytics instead.
      tags: [Portfolio]
      responses:
        '200':
          description: Gross/net exposure, leverage ratio, and largest-position concentration.
          content:
            application/json:
              schema:
                type: object
                properties:
                  grossExposure:
                    type: number
                  netExposure:
                    type: number
                  leverageRatio:
                    type: number
                  largestPositionPct:
                    type: number
                  missingPriceSymbols:
                    type: array
                    items:
                      type: string
                required: [grossExposure, netExposure, leverageRatio, largestPositionPct, missingPriceSymbols]
        '401':
          description: Missing or invalid bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |-
        Every proxied service verifies this itself against its own JWT
        secret; the gateway's own requireAuth (used only by the 10
        in-process routes) uses the identical verifyAccessToken /
        InvalidTokenError contract from @tradosphere/auth as every other
        service.
  schemas:
    ErrorResponse:
      type: object
      description: |-
        The generic shape every service's setErrorHandler falls back to for
        unexpected errors, and the shape most caught domain errors use
        (sometimes with extra fields alongside error, e.g. 409 responses
        that also carry missingPriceSymbols). Not a Fastify-default
        {statusCode,error,message} envelope -- just {error, ...}.
      properties:
        error:
          type: string
      required: [error]
    ValidationFailure:
      type: object
      description: The zod-failure shape validateBody() returns on every 400, identical across every service (reused verbatim from services/education's original).
      properties:
        error:
          type: string
          example: Validation failed
        details:
          type: array
          items:
            type: object
            properties:
              path:
                type: string
              message:
                type: string
            required: [path, message]
      required: [error, details]
...
    PositionDirection:
      type: string
      enum: [long, short]
    Position:
      type: object
      description: Net position for one symbol, computed from open trade records only as a quantity-weighted average entry price (Decision D17).
      properties:
        symbol:
          type: string
        direction:
          $ref: '#/components/schemas/PositionDirection'
        quantity:
          type: number
        averageEntryPrice:
          type: number
      required: [symbol, direction, quantity, averageEntryPrice]
    PortfolioSnapshot:
      type: object
      description: One persisted row from POST /v1/portfolio/snapshot. totalEquity = startingCash + realizedPnl + unrealizedPnl, which is algebraically identical to cashBalance + positionsValue (Decision D17).
      properties:
        id:
          type: string
        userId:
          type: string
        cashBalance:
          type: number
        positionsValue:
          type: number
        realizedPnl:
          type: number
        unrealizedPnl:
          type: number
        totalEquity:
          type: number
        label:
          type: string
        asOf:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, cashBalance, positionsValue, realizedPnl, unrealizedPnl, totalEquity, asOf, createdAt]
    CreateSnapshotInput:
      type: object
      description: Every field is optional. Deliberately excludes every computed figure (cashBalance, positionsValue, realizedPnl, unrealizedPnl, totalEquity) -- a snapshot can never be fabricated or tampered with (Forge charter rule 2).
      properties:
        label:
          type: string
          minLength: 1
          maxLength: 200
        asOf:
          type: string
          format: date-time
          description: Defaults to the current time when omitted.
    PerformanceMetrics:
      type: object
      description: Portfolio-level performance totals only (Decision D17) -- see GET /v1/portfolio/performance's description for what's deliberately out of scope.
      properties:
        startingCash:
          type: number
        totalEquity:
          type: number
        totalReturn:
          type: number
        totalReturnPct:
          type: number
        realizedPnl:
          type: number
        unrealizedPnl:
          type: number
      required: [startingCash, totalEquity, totalReturn, totalReturnPct, realizedPnl, unrealizedPnl]
    AllocationEntry:
      type: object
      properties:
        symbol:
          type: string
        direction:
          $ref: '#/components/schemas/PositionDirection'
        marketValue:
          type: number
        allocationPct:
          type: number
      required: [symbol, direction, marketValue, allocationPct]
...